requirements.txt
flask==3.0.2
python-dotenv==1.0.1
spotipy==2.23.0
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0
requests==2.31.0
python-multipart==0.0.9
flask-cors==4.0.0
gunicorn==21.2.0
huggingface-hub==0.20.3 